+++ features/cljs-time/src/nbb/impl/cljs_time.cljs
@@ -24,12 +24,17 @@ Remaining fns are used by the community"
    'days (copy-var t/days core-ns)
    'today (copy-var t/today core-ns)
    'minus (copy-var t/minus core-ns)
-   'plus (copy-var t/plus core-ns)})
+   'plus (copy-var t/plus core-ns)
+   'to-default-time-zone (copy-var t/to-default-time-zone core-ns)
+   'local-date (copy-var t/local-date core-ns)})
 
 (def coerce-namespace
   "Only to-long is used by logseq.graph-parser"
   {'to-long (copy-var tc/to-long coerce-ns)
-   'from-long (copy-var tc/from-long coerce-ns)})
+   'from-long (copy-var tc/from-long coerce-ns)
+   'to-local-date (copy-var tc/to-local-date coerce-ns)
+   'to-date (copy-var tc/to-date coerce-ns)
+   'to-local-date-time (copy-var tc/to-local-date-time coerce-ns)})
 
 (def format-namespace
   "All fns in this ns are used by logseq.graph-parser"